requirements.txt
absl-py
aiohttp
aiohttp-retry
aiosignal
anyio
argon2-cffi
argon2-cffi-bindings
arrow
asgiref
asttokens
astunparse
async-lru
attrs
Babel
beautifulsoup4
bleach
blinker
cachetools
certifi
cffi
charset-normalizer
click
colorama
comm
comtypes
contourpy
crispy-bootstrap5
cryptography
cycler
debugpy
decorator
defusedxml
distlib
dj-database-url
Django
django-allauth
django-cors-headers
django-crispy-forms
django-email
django-environ
django-filter
django-model-utils
django-rest-knox
django-tailwind
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djoser
EasyProcess
ecapture
entrypoint2
executing
fastjsonschema
filelock
Flask
flatbuffers
fonttools
fqdn
frozenlist
fuzzywuzzy
gast
google-auth
google-auth-oauthlib
google-pasta
greenlet
grpcio
h11
h5py
httpcore
httpx
idna
install
ipykernel
ipython
ipywidgets
isoduration
itsdangerous
jaraco.context
jedi
Jinja2
joblib
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
keras
kiwisolver
libclang
Markdown
MarkupSafe
matplotlib
matplotlib-inline
mistune
ml-dtypes
more-itertools
MouseInfo
mss
multidict
mysql-connector-python
nbclient
nbconvert
nbformat
nest-asyncio
notebook
notebook_shim
numpy
oauthlib
opencv-python
opt-einsum
overrides
package-name
packaging
pandas
pandocfilters
parso
Pillow
pipenv
platformdirs
prometheus_client
prompt-toolkit
protobuf
psutil
psycopg2
psycopg2-binary
pure-eval
pyasn1
pyasn1-modules
PyAutoGUI
pycparser
PyGetWindow
Pygments
pyjokes
PyJWT
PyMsgBox
PyMySQL
pyparsing
pyperclip
pypiwin32
pypng
PyRect
pyscreenshot
PyScreeze
python-dateutil
python-decouple
python-json-logger
python3-openid
pyttsx3
pytweening
pytz
pywin32
pywinpty
PyYAML
pyzbar
pyzmq
qrcode
qtconsole
QtPy
referencing
requests
requests-oauthlib
rfc3339-validator
rfc3986-validator
rpds-py
rsa
scikit-learn
scipy
Send2Trash
six
sniffio
social-auth-app-django
social-auth-core
soupsieve
SpeechRecognition
SQLAlchemy
sqlparse
stack-data
stripe
tensorboard
tensorboard-data-server
tensorflow-estimator
tensorflow-io-gcs-filesystem
termcolor
terminado
threadpoolctl
tinycss2
tornado
traitlets
twilio
types-python-dateutil
typing_extensions
tzdata
uri-template
urllib3
virtualenv
wcwidth
webcolors
webencodings
websocket-client
Werkzeug
widgetsnbextension
wikipedia
wolframalpha
wrapt
xmltodict
yarl